+++ src/account/AccountId.js
@@ -160,4 +160,15 @@ export default class AccountId {
         id._checksum = this._checksum;
         return id;
     }
+
+    /**
+     * @param {AccountId} other
+     * @returns {number}
+     */
+    compare(other) {
+        return entity_id.compare(
+            [this.shard, this.realm, this.num],
+            [other.shard, other.realm, other.num]
+        );
+    }
 }