+++ src/file/FileId.js
@@ -133,4 +133,15 @@ export default class FileId {
         id._checksum = this._checksum;
         return id;
     }
+
+    /**
+     * @param {FileId} other
+     * @returns {number}
+     */
+    compare(other) {
+        return entity_id.compare(
+            [this.shard, this.realm, this.num],
+            [other.shard, other.realm, other.num]
+        );
+    }
 }