+++ src/token/TokenId.js
@@ -148,4 +148,15 @@ export default class TokenId {
         id._checksum = this._checksum;
         return id;
     }
+
+    /**
+     * @param {TokenId} other
+     * @returns {number}
+     */
+    compare(other) {
+        return entity_id.compare(
+            [this.shard, this.realm, this.num],
+            [other.shard, other.realm, other.num]
+        );
+    }
 }