+++ src/topic/TopicId.js
@@ -137,4 +137,15 @@ export default class TopicId {
         id._checksum = this._checksum;
         return id;
     }
+
+    /**
+     * @param {TopicId} other
+     * @returns {number}
+     */
+    compare(other) {
+        return entity_id.compare(
+            [this.shard, this.realm, this.num],
+            [other.shard, other.realm, other.num]
+        );
+    }
 }